Simplified builders and excluded from code analysis :sparkles:
