+++ Source/Tag/ID32v2TagBuilder.swift
@@ -13,13 +13,309 @@ import Foundation
  It let the user set all the frames included in ID32TagCommonStringFramesBuilder,
  ID32TagCommonCompositeFramesBuilder and RecordingHourMinuteDayMonthYearFramesBuilder.
  */
-public class ID32v2TagBuilder: ID32TagCommonCompositeFramesBuilder,
-                               TagBuilder,
-                               RecordingHourMinuteDayMonthYearFramesBuilder {
+public class ID32v2TagBuilder: TagBuilder {
+    var frames: [FrameName: ID3Frame] = [:]
+
     /**
      Init a ID32v2TagBuilder instance.
      */
-    override public init() {}
+    public init() {}
+
+    /**
+      Set the title frame to be written by ID3TagEditor.
+     
+      - parameter frame: the title frame as a ID3FrameWithStringContent instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func title(frame: ID3FrameWithStringContent) -> Self {
+        frames[.title] = frame
+        return self
+    }
+
+    /**
+      Set the album frame to be written by ID3TagEditor.
+     
+      - parameter frame: the album frame as a ID3FrameWithStringContent instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func album(frame: ID3FrameWithStringContent) -> Self {
+        frames[.album] = frame
+        return self
+    }
+
+    /**
+      Set the albumArtist frame to be written by ID3TagEditor.
+     
+      - parameter frame: the albumArtist frame as a ID3FrameWithStringContent instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func albumArtist(frame: ID3FrameWithStringContent) -> Self {
+        frames[.albumArtist] = frame
+        return self
+    }
+
+    /**
+      Set the artist frame to be written by ID3TagEditor.
+     
+      - parameter frame: the artist frame as a ID3FrameWithStringContent instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func artist(frame: ID3FrameWithStringContent) -> Self {
+        frames[.artist] = frame
+        return self
+    }
+
+    /**
+      Set the composer frame to be written by ID3TagEditor.
+     
+      - parameter frame: the composer frame as a ID3FrameWithStringContent instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func composer(frame: ID3FrameWithStringContent) -> Self {
+        frames[.composer] = frame
+        return self
+    }
+
+    /**
+      Set the conductor frame to be written by ID3TagEditor.
+     
+      - parameter frame: the conductor frame as a ID3FrameWithStringContent instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func conductor(frame: ID3FrameWithStringContent) -> Self {
+        frames[.conductor] = frame
+        return self
+    }
+
+    /**
+      Set the contentGrouping frame to be written by ID3TagEditor.
+     
+      - parameter frame: the contentGrouping frame as a ID3FrameWithStringContent instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func contentGrouping(frame: ID3FrameWithStringContent) -> Self {
+        frames[.contentGrouping] = frame
+        return self
+    }
+
+    /**
+      Set the copyright frame to be written by ID3TagEditor.
+     
+      - parameter frame: the copyright frame as a ID3FrameWithStringContent instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func copyright(frame: ID3FrameWithStringContent) -> Self {
+        frames[.copyright] = frame
+        return self
+    }
+
+    /**
+      Set the encodedBy frame to be written by ID3TagEditor.
+     
+      - parameter frame: the encodedBy frame as a ID3FrameWithStringContent instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func encodedBy(frame: ID3FrameWithStringContent) -> Self {
+        frames[.encodedBy] = frame
+        return self
+    }
+
+    /**
+      Set the encoderSettings frame to be written by ID3TagEditor.
+     
+      - parameter frame: the encoderSettings frame as a ID3FrameWithStringContent instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func encoderSettings(frame: ID3FrameWithStringContent) -> Self {
+        frames[.encoderSettings] = frame
+        return self
+    }
+
+    /**
+      Set the lyricist frame to be written by ID3TagEditor.
+     
+      - parameter frame: the lyricist frame as a ID3FrameWithStringContent instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func lyricist(frame: ID3FrameWithStringContent) -> Self {
+        frames[.lyricist] = frame
+        return self
+    }
+
+    /**
+      Set the mixArtist frame to be written by ID3TagEditor.
+     
+      - parameter frame: the mixArtist frame as a ID3FrameWithStringContent instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func mixArtist(frame: ID3FrameWithStringContent) -> Self {
+        frames[.mixArtist] = frame
+        return self
+    }
+
+    /**
+      Set the publisher frame to be written by ID3TagEditor.
+     
+      - parameter frame: the publisher frame as a ID3FrameWithStringContent instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func publisher(frame: ID3FrameWithStringContent) -> Self {
+        frames[.publisher] = frame
+        return self
+    }
+
+    /**
+      Set the subtitle frame to be written by ID3TagEditor.
+     
+      - parameter frame: the subtitle frame as a ID3FrameWithStringContent instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func subtitle(frame: ID3FrameWithStringContent) -> Self {
+        frames[.subtitle] = frame
+        return self
+    }
+
+    /**
+      Set the beats per minute frame to be written by ID3TagEditor.
+     
+      - parameter frame: the beats per minute frame as a ID3FrameWithIntegerContent instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func beatsPerMinute(frame: ID3FrameWithIntegerContent) -> Self {
+        frames[.beatsPerMinute] = frame
+        return self
+    }
+
+    /**
+      Set the genre frame to be written by ID3TagEditor.
+     
+      - parameter frame: the genre frame as a ID3FrameGenre instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func genre(frame: ID3FrameGenre) -> Self {
+        frames[.genre] = frame
+        return self
+    }
+
+    /**
+      Set the discPosition frame to be written by ID3TagEditor.
+     
+      - parameter frame: the discPosition frame as a ID3FramePartOfTotal instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func discPosition(frame: ID3FramePartOfTotal) -> Self {
+        frames[.discPosition] = frame
+        return self
+    }
+
+    /**
+      Set the trackPosition frame to be written by ID3TagEditor.
+     
+      - parameter frame: the trackPosition frame as a ID3FramePartOfTotal instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func trackPosition(frame: ID3FramePartOfTotal) -> Self {
+        frames[.trackPosition] = frame
+        return self
+    }
+
+    /**
+     Set an attachedPicture frame to be written by ID3TagEditor. You can set multiple
+     attached picture frames per tag, one for each of the picture types contained in ID3PictureType.
+     
+      - parameter pictureType: the ID3PictureType associated to the attachedPicture frame to be set.
+      - parameter frame: the attachedPicture frame as a ID3FrameAttachedPicture instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func attachedPicture(pictureType: ID3PictureType, frame: ID3FrameAttachedPicture) -> Self {
+        frames[.attachedPicture(pictureType)] = frame
+        return self
+    }
+
+    /**
+     Set an unsynchronisedLyrics frame to be written by ID3TagEditor. You can set multiple
+     unsynchronised lyrics frames per tag, one for each of the languages contained in ID3FrameContentLanguage.
+     
+      - parameter language: the ID3FrameContentLanguage associated to the unsynchronisedLyrics frame to be set.
+      - parameter frame: the unsynchronisedLyrics frame as a ID3FrameWithLocalizedContent instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func unsynchronisedLyrics(language: ID3FrameContentLanguage,
+                                     frame: ID3FrameWithLocalizedContent) -> Self {
+        frames[.unsynchronizedLyrics(language)] = frame
+        return self
+    }
+
+    /**
+     Set an comment frame to be written by ID3TagEditor. You can set multiple
+     comment frames per tag, one for each of the languages contained in ID3FrameContentLanguage.
+     
+      - parameter language: the ID3FrameContentLanguage associated to the comment frame to be set.
+      - parameter frame: the comment frame as a ID3FrameWithLocalizedContent instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func comment(language: ID3FrameContentLanguage,
+                        frame: ID3FrameWithLocalizedContent) -> Self {
+        frames[.comment(language)] = frame
+        return self
+    }
+
+    /**
+      Set the recordingDayMonth frame to be written by ID3TagEditor.
+     
+      - parameter frame: the recordingDayMonth frame as a ID3FrameWithStringContent instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func recordingDayMonth(frame: ID3FrameRecordingDayMonth) -> Self {
+        frames[.recordingDayMonth] = frame
+        return self
+    }
+
+    /**
+      Set the recordingHourMinute frame to be written by ID3TagEditor.
+     
+      - parameter frame: the recordingHourMinute frame as a ID3FrameWithStringContent instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func recordingHourMinute(frame: ID3FrameRecordingHourMinute) -> Self {
+        frames[.recordingHourMinute] = frame
+        return self
+    }
+
+    /**
+      Set the recordingYear frame to be written by ID3TagEditor.
+     
+      - parameter frame: the recordingYear frame as a ID3FrameWithStringContent instance.
+     
+      - returns: the instance of the builder.
+     */
+    public func recordingYear(frame: ID3FrameWithIntegerContent) -> Self {
+        frames[.recordingYear] = frame
+        return self
+    }
 
     /**
      Build and ID3Tag version 2.
